+++ pkg/router/router.go
@@ -23,12 +23,14 @@ func Setup() *gin.Engine {
 	usersMap := map[string]string{
 		config.Server.Username: config.Server.Password,
 	}
+
 	authorized := r.Group("/", gin.BasicAuth(usersMap))
 	logins := authorized.Group("/logins")
 	{
 		logins.GET("/", login.GetLogins)
 		logins.GET("/:id", login.GetLogin)
 		logins.POST("/", login.CreateLogin)
+		logins.POST("/:action", login.PostHandler)
 		logins.PUT("/:id", login.UpdateLogin)
 		logins.DELETE("/:id", login.DeleteLogin)
 	}